--- a/apps/ledger-live-mobile/src/components/RootNavigator/AccountsNavigator.tsx
+++ b/apps/ledger-live-mobile/src/components/RootNavigator/AccountsNavigator.tsx
@@ -1,28 +1,29 @@
-import React, { useMemo } from "react";
+import React, { useMemo, lazy } from "react";
 import { createStackNavigator } from "@react-navigation/stack";
 import { useTheme } from "styled-components/native";
 import { useSelector } from "react-redux";
 import { readOnlyModeEnabledSelector } from "../../reducers/settings";
 import { ScreenName } from "../../const";
-import Accounts from "../../screens/Accounts";
-import Account from "../../screens/Account";
-import NftCollection from "../../screens/Nft/NftCollection";
-import NftGallery from "../../screens/Nft/NftGallery";
-import NftViewer from "../Nft/NftViewer";
-import NftCollectionHeaderTitle from "../../screens/Nft/NftCollection/NftCollectionHeaderTitle";
-import NftGalleryHeaderTitle from "../../screens/Nft/NftGallery/NftGalleryHeaderTitle";
 import { getStackNavigatorConfig } from "../../navigation/navigatorConfig";
-import ReadOnlyAccounts from "../../screens/Accounts/ReadOnly/ReadOnlyAccounts";
-import ReadOnlyAssets from "../../screens/Portfolio/ReadOnlyAssets";
-
-import Asset from "../../screens/WalletCentricAsset";
-import ReadOnlyAsset from "../../screens/WalletCentricAsset/ReadOnly";
-import Assets from "../../screens/Assets";
-
-import ReadOnlyAccount from "../../screens/Account/ReadOnly/ReadOnlyAccount";
-
 import type { AccountsNavigatorParamList } from "./types/AccountsNavigator";
 import { hasNoAccountsSelector } from "../../reducers/accounts";
+import { withSuspense } from "~/helpers/withSuspense";
+
+
+const Accounts = lazy(() => import("../../screens/Accounts"));
+const Account = lazy(() => import("../../screens/Account"));
+const NftCollection = lazy(() => import("../../screens/Nft/NftCollection"));
+const NftGallery = lazy(() => import("../../screens/Nft/NftGallery"));
+const NftViewer = lazy(() => import("../Nft/NftViewer"));
+const NftCollectionHeaderTitle = lazy(() => import("../../screens/Nft/NftCollection/NftCollectionHeaderTitle"));
+const NftGalleryHeaderTitle = lazy(() => import("../../screens/Nft/NftGallery/NftGalleryHeaderTitle"));
+const ReadOnlyAccounts = lazy(() => import("../../screens/Accounts/ReadOnly/ReadOnlyAccounts"));
+const ReadOnlyAssets = lazy(() => import("../../screens/Portfolio/ReadOnlyAssets"));
+const Asset = lazy(() => import("../../screens/WalletCentricAsset"));
+const ReadOnlyAsset = lazy(() => import("../../screens/WalletCentricAsset/ReadOnly"));
+const Assets = lazy(() => import("../../screens/Assets"));
+const ReadOnlyAccount = lazy(() => import("../../screens/Account/ReadOnly/ReadOnlyAccount"));
+
 
 const Stack = createStackNavigator<AccountsNavigatorParamList>();
 
@@ -37,47 +38,47 @@ export default function AccountsNavigator() {
     <Stack.Navigator screenOptions={stackNavConfig}>
       <Stack.Screen
         name={ScreenName.Accounts}
-        component={readOnlyModeEnabled ? ReadOnlyAccounts : Accounts}
+        component={withSuspense(readOnlyModeEnabled ? ReadOnlyAccounts : Accounts)}
         options={{
           headerShown: false,
         }}
       />
       <Stack.Screen
         name={ScreenName.Account}
-        component={readOnlyModeEnabled ? ReadOnlyAccount : Account}
+        component={withSuspense(readOnlyModeEnabled ? ReadOnlyAccount : Account)}
         options={{ headerShown: false }}
       />
       <Stack.Screen
         name={ScreenName.NftCollection}
-        component={NftCollection}
+        component={withSuspense(NftCollection)}
         options={{
           headerTitle: () => <NftCollectionHeaderTitle />,
         }}
       />
       <Stack.Screen
         name={ScreenName.NftGallery}
-        component={NftGallery}
+        component={withSuspense(NftGallery)}
         options={{
           headerTitle: () => <NftGalleryHeaderTitle />,
         }}
       />
       <Stack.Screen
         name={ScreenName.NftViewer}
-        component={NftViewer}
+        component={withSuspense(NftViewer)}
         options={{
           headerTitle: "",
         }}
       />
       <Stack.Screen
         name={ScreenName.Assets}
-        component={readOnlyModeEnabled ? ReadOnlyAssets : Assets}
+        component={withSuspense(readOnlyModeEnabled ? ReadOnlyAssets : Assets)}
         options={{
           headerShown: false,
         }}
       />
       <Stack.Screen
         name={ScreenName.Asset}
-        component={readOnlyModeEnabled ? ReadOnlyAsset : Asset}
+        component={withSuspense(readOnlyModeEnabled ? ReadOnlyAsset : Asset)}
         options={{
           headerShown: false,
         }}