--- a/apps/ledger-live-mobile/tsconfig.json
+++ b/apps/ledger-live-mobile/tsconfig.json
@@ -4,19 +4,43 @@
     "allowJs": false,
     "checkJs": false,
     "jsx": "react-native",
-    "lib": ["es2017", "DOM"],
-    "types": ["jest", "@testing-library/jest-native/extend-expect"],
+    "lib": [
+      "es2017",
+      "DOM"
+    ],
+    "types": [
+      "jest",
+      "@testing-library/jest-native/extend-expect"
+    ],
     "noEmit": true,
+    "module": "es2020",
     "target": "esnext",
     "incremental": true,
     "baseUrl": ".",
     "rootDir": ".",
     "paths": {
-      "~/*": ["src/*"],
-      "@utils/*": ["./src/utils/*"],
-      "@tests/*": ["src/__tests__/*"]
+      "~/*": [
+        "src/*"
+      ],
+      "@utils/*": [
+        "./src/utils/*"
+      ],
+      "@tests/*": [
+        "src/__tests__/*"
+      ]
     }
   },
-  "include": ["src/**/*", "e2e/**/*", "**/*.d.ts", "services/**/*"],
-  "exclude": ["node_modules", "babel.config.js", "metro.config.js", "jest.config.js", "**/*.js"]
+  "include": [
+    "src/**/*",
+    "e2e/**/*",
+    "**/*.d.ts",
+    "services/**/*"
+  ],
+  "exclude": [
+    "node_modules",
+    "babel.config.js",
+    "metro.config.js",
+    "jest.config.js",
+    "**/*.js"
+  ]
 }