<!--
Thank you for your contribution! 👍
Please make sure to read CONTRIBUTING.md if you have not already. Pull Requests that do not comply with the rules will be arbitrarily closed.
-->

### 📝 Description

In order for the app to open faster, we tried to lazy load some screens on LLM.
The same was done for [LLD](https://github.com/LedgerHQ/ledger-live/pull/5072) and it resulted in significant performance gains.

In order to measure the time for the app to start, we use the `appTimeToInteractiveMilliseconds` property from the `StartupTimeMarker` component which is accessible from the Settings > Debug > Performance.

After a first try where we lazy load about 35 screens in LLM, unfortunately we don't notice any improvement in the app start of LLM.

### BEFORE

#### DEV :

 1 - 22164 ms
 2 - 21793 ms
 3 - 21879 ms
 4 - 22114 ms
 5 - 22059 ms

Average : 22002 ms (22 seconds)

#### STAGING :

 1 -  2960 ms
 2 - 2917 ms
 3 - 3122 ms
 4 - 3031  ms
 5 - 3228 ms
 6 - 3230 ms
 7 - 3457 ms
 8 - 3262 ms
 9 - 3130 ms
10 - 3141 ms

Average : 3148 ms (3,1 seconds)

### ❓ Context

- **JIRA or GitHub link**: [LIVE-9814]<!-- Attach the relevant ticket number if applicable. (e.g., JIRA-123 or #123) -->

### ✅ Checklist

Pull Requests must pass the CI and be code reviewed. Set as Draft if the PR is not ready.

- [ ] `npx changeset` was attached.
- [ ] **Covered by automatic tests.** <!-- if not, please explain. (Feature must be tested / Bugfix must bring non-regression) -->
- [ ] **Impact of the changes:** <!-- Please take some time to list the impact & what specific areas Quality Assurance (QA) should focus on -->
  - 

---

### 🧐 Checklist for the PR Reviewers

<!-- Please do not edit this if you are the PR author -->

- [ ] **The code aligns with the requirements** described in the linked JIRA or GitHub issue.
- [ ] **The PR description clearly documents the changes** made and explains any technical trade-offs or design decisions.
- [ ] **There are no undocumented trade-offs**, technical debt, or maintainability issues.
- [ ] **The PR has been tested** thoroughly, and any potential edge cases have been considered and handled.
- [ ] **Any new dependencies** have been justified and documented.
- [ ] **Performance** considerations have been taken into account. (changes have been profiled or benchmarked if necessary)


[LIVE-9814]: https://ledgerhq.atlassian.net/browse/LIVE-9814?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ